+++ apps/cms-v2/src/modules/bulk-sync/bulk-sync-processor.ts
@@ -1,8 +1,8 @@
 import { BulkImportProductFragment } from "../../../generated/graphql";
 import { AnyProviderConfigSchemaType } from "../configuration";
-import { ContentfulBulkSyncProcessor } from "../contentful/contentful-bulk-sync-processor";
-import { DatocmsBulkSyncProcessor } from "../datocms/datocms-bulk-sync-processor";
-import { StrapiBulkSyncProcessor } from "../strapi/strapi-bulk-sync-processor";
+import { ContentfulBulkSyncProcessor } from "../providers/contentful/contentful-bulk-sync-processor";
+import { DatocmsBulkSyncProcessor } from "../providers/datocms/datocms-bulk-sync-processor";
+import { StrapiBulkSyncProcessor } from "../providers/strapi/strapi-bulk-sync-processor";
 
 export type BulkSyncProcessorHooks = {
   onUploadStart?: (context: { variantId: string }) => void;
@@ -11,9 +11,13 @@ export type BulkSyncProcessorHooks = {
 };
 
 export interface BulkSyncProcessor {
-  uploadProducts(products: BulkImportProductFragment[], hooks: Hooks): Promise<void>;
+  uploadProducts(
+    products: BulkImportProductFragment[],
+    hooks: BulkSyncProcessorHooks
+  ): Promise<void>;
 }
 
+// todo extract to shared
 export const BulkSyncProcessorFactory = {
   create(config: AnyProviderConfigSchemaType): BulkSyncProcessor {
     switch (config.type) {
