+++ apps/cms-v2/src/modules/providers/datocms/datocms-client.ts
@@ -1,6 +1,6 @@
 import { buildClient, Client, SimpleSchemaTypes } from "@datocms/cma-client-node";
-import { WebhookProductVariantFragment } from "../../../generated/graphql";
-import { DatocmsProviderConfigType } from "../configuration/schemas/datocms-provider.schema";
+import { WebhookProductVariantFragment } from "../../../../generated/graphql";
+import { DatocmsProviderConfigType } from "../../configuration/schemas/datocms-provider.schema";
 
 type Context = {
   configuration: DatocmsProviderConfigType;
