+++ apps/cms-v2/src/modules/providers/datocms/datocms-config-form.tsx
@@ -4,10 +4,10 @@ import { Input, Select } from "@saleor/react-hook-form-macaw";
 import { useRouter } from "next/router";
 import React, { useEffect, useMemo } from "react";
 import { useForm } from "react-hook-form";
-import { printSaleorProductFields } from "../configuration/print-saleor-product-fields";
-import { DatocmsProviderConfigInputType } from "../configuration/schemas/datocms-provider.schema";
-import { trpcClient } from "../trpc/trpc-client";
-import { ButtonsBox } from "../ui/buttons-box";
+import { printSaleorProductFields } from "../../configuration/print-saleor-product-fields";
+import { DatocmsProviderConfigInputType } from "../../configuration/schemas/datocms-provider.schema";
+import { trpcClient } from "../../trpc/trpc-client";
+import { ButtonsBox } from "../../ui/buttons-box";
 
 type FormShape = Omit<DatocmsProviderConfigInputType, "type">;
