+++ apps/cms-v2/src/modules/providers/strapi/strapi.router.ts
@@ -1,0 +1,3 @@
+import { router } from "../../trpc/trpc-server";
+
+export const strapiRouter = router({});
