--- apps/cms-v2/src/modules/strapi/strapi.router.ts
@@ -1,3 +1,0 @@
-import { router } from "../trpc/trpc-server";
-
-export const strapiRouter = router({});
