+++ apps/cms-v2/src/modules/trpc/trpc-app-router.ts
@@ -1,8 +1,8 @@
 import { channelProviderConnectionRouter } from "../channel-provider-connection/channel-provider-connection.router";
-import { contentfulRouter } from "../contentful/contentful.router";
-import { datocmsRouter } from "../datocms/datocms.router";
+import { contentfulRouter } from "../providers/contentful/contentful.router";
+import { datocmsRouter } from "../providers/datocms/datocms.router";
 import { providersListRouter } from "../providers-listing/providers-list.router";
-import { strapiRouter } from "../strapi/strapi.router";
+import { strapiRouter } from "../providers/strapi/strapi.router";
 import { router } from "./trpc-server";
 
 export const appRouter = router({
